# feat(stdlib): improve filter performance when filtering on values