--- a/internal/arrowutil/gen.go
+++ b/internal/arrowutil/gen.go
@@ -4,3 +4,4 @@ package arrowutil
 //go:generate tmpl -data=@types.tmpldata -o builder.gen.go builder.gen.go.tmpl
 //go:generate tmpl -data=@types.tmpldata -o iterator.gen.go iterator.gen.go.tmpl
 //go:generate tmpl -data=@types.tmpldata -o iterator.gen_test.go iterator.gen_test.go.tmpl
+//go:generate tmpl -data=@types.tmpldata -o filter.gen.go filter.gen.go.tmpl