This improves filter's performance when the function will evaluate
values that are not part of the group key.

Related to #2211.

### Done checklist
- [x] docs/SPEC.md updated
- [x] Test cases written