@jsternberg 2019-12-04
Here's a performance difference:

```
jsternberg:[~/g/s/g/i/flux] $ (feat/optimize-filter-values $=) benchcmp old.txt new.txt
benchmark                          old ns/op       new ns/op       delta
BenchmarkFilter_Values/1000-12     73667911676     59678115452     -18.99%

benchmark                          old allocs     new allocs     delta
BenchmarkFilter_Values/1000-12     1089074700     905505262      -16.86%

benchmark                          old bytes       new bytes       delta
BenchmarkFilter_Values/1000-12     29404196856     21521477224     -26.81%
```

The main bulk of the slowness appears to be when calling the function so this change primarily gets rid of the problems with filter itself.